Add component source for sidebar menu.
